+++ packages/adk-cli/src/server/controllers/health.controller.ts
@@ -1,0 +1,7 @@
+import type { Context } from "hono";
+
+export class HealthController {
+	public async getHealth(c: Context) {
+		return c.json({ status: "ok", timestamp: new Date().toISOString() });
+	}
+}
