+++ packages/adk-cli/src/server/controllers/index.ts
@@ -1,0 +1,2 @@
+export * from "./health.controller.js";
+export * from "./agent.controller.js";
