+++ packages/adk-cli/src/server/handlers/index.ts
@@ -1,0 +1,1 @@
+export * from "./socket.handler.js";
