+++ packages/adk-cli/src/server/index.ts
@@ -1,0 +1,10 @@
+// Main server class - the only export needed outside the server module
+export { ADKServer } from "./adk-server.js";
+
+// Internal exports for server module organization
+// These are used internally between server components but not exported publicly
+// export * from "./types.js";
+// export * from "./services/index.js";
+// export * from "./controllers/index.js";
+// export * from "./middleware/index.js";
+// export * from "./handlers/index.js";
