+++ packages/adk-cli/src/server/middleware/cors.middleware.ts
@@ -1,0 +1,6 @@
+import { cors } from "hono/cors";
+
+export const corsMiddleware = cors({
+	origin: ["http://localhost:3000", "http://localhost:3001"],
+	credentials: true,
+});
